https://github.com/WWBN/AVideo/issues/6954#issuecomment-1234921484

+++ plugin/ReportVideo/Objects/videos_reported.php
@@ -141,7 +141,7 @@ class VideosReported extends ObjectYPT {
         }
 
         $saved = parent::save();
-        if($saved){
+        if($saved && class_exists('Cache')){
             Cache::deleteFirstPageCache();
         }
         return $saved;
